login and registration working stillhave to fix table not showing for authenticated users also fix links and add functionality for file upload

+++ bookLibrary.php
@@ -14,29 +14,33 @@ $books = $cmd->fetchAll();
 
 // echo out the table to the site
     echo'<h2>Book Library</h2>';
-    echo '<table><thead><th>Book Title</th><th>Book Author</th><th>Publish Year</th><th>Book Genre</th><th>Book Publisher</th><th>Made Into Movie</th><th>Streaming Service</th><th>Actions</tr></thead><tbody>';
+    echo '<table><thead><th>Book Title</th><th>Book Author</th><th>Publish Year</th><th>Book Genre</th><th>Book Publisher</th><th>Made Into Movie</th><th>Streaming Service</th>';
+    if (!empty(currentSession['username'])) {
+        echo'<th>Actions</th>';
+    }
+    echo '</thead>';
 
-    foreach ($books as $books){ 
-        echo '<tr>
-            <td>'.$books['bookTitle'].'</td>
-            <td>'.$books['bookAuthor'].'</td>
-            <td>'.$books['publishYear'].'</td>
-            <td>'.$books['bookGenre'].'</td>
-            <td>'.$books['bookPublisher'].'</td>
-            <td>'.$books['madeIntoMovie'].'</td>
-            <td>'.$books['streamingService'].'</td>
-            <td>
+foreach ($books as $books){ 
+    echo '<tr>
+        <td>'.$books['bookTitle'].'</td>
+        <td>'.$books['bookAuthor'].'</td>
+        <td>'.$books['publishYear'].'</td>
+        <td>'.$books['bookGenre'].'</td>
+        <td>'.$books['bookPublisher'].'</td>
+        <td>'.$books['madeIntoMovie'].'</td>
+        <td>'.$books['streamingService'].'</td>';
+        if (!empty($currentSession['username'])){    
+            echo '<td class="actions">    
                 <a href="edit-book.php?bookId=' . $books['bookId'] . '">
                     Edit
                 </a>&nbsp;
                 <a href="delete-book.php?bookId=' .$books['bookId'] . '" onclick="return confirmDelete();">
                     Delete
                 </a>
-            </td>
-            </tr>';
+            </td>';
         }
-
-    echo '</table>';
+        echo '</tr>';
+}
 
 $database = null;
 ?>
